refactor: co-locate test files with respective logic files
